+++ webapp/polls/urls.py
@@ -9,13 +9,11 @@ urlpatterns = [
     path('', views.IndexView.as_view(), name='index'),
 
     # ex: /polls/5/
-    path('<int:question_id>/', views.detail, name='detail'),
-    # path('<int:question_id>/', views.DetailView.as_view(), name='detail'),
+    path('<int:pk>/', views.DetailView.as_view(), name='detail'),
 
     # ex: /polls/5/vote/
     path('<int:question_id>/vote/', views.vote, name='vote'),
 
     # ex: /polls/5/results/
-    path('<int:question_id>/results/', views.results, name='results'),
-    # path('<int:question_id>/results/', views.ResultsView.as_view(), name='results'),
+    path('<int:pk>/results/', views.ResultsView.as_view(), name='results'),
 ]